requirements.txt
tensorflow
numpy
pyyaml
h5py
easydict
tqdm